dynamic scenes
made "list of buttons" and "upgrade button" as their own scenes so they can be generated repeatedly.

"list of buttons" is manually instantiated into the main scene and will display buttons added to it in code.

"singular_upgrade_container" is a model/template for coding the upgrades that will be looped over and added into the "list_of_upgrade_buttons" scene standalone

- easier to code as we don't have to care about the paths
- cleaner editor scene tree
- good for future physical changes to the upgrade style - just updating template and possibly adding appropriate code